refactor(): reorganize file structure page & components
