+++ src/pages/Home/Home.tsx
@@ -3,8 +3,8 @@
 import { useEffect, useState, useCallback } from "react"
 import { useUserStore } from "@/store/useUserStore"
 import { UiButton, UiLoading } from "@/components/ui/index"
-import TaskAddModal from "@/components/task/TaskAddModal"
-import TaskColumn from "@/components/task/TaskColumn"
+import TaskAddModal from "@/pages/Home/components/TaskAddModal"
+import TaskColumn from "@/pages/Home/components/TaskColumn"
 import { statusItems } from "@/constants/options"
 import { getTasks, addTask } from "@/api/tasks"
 import { fakeDelay, nowDate } from "@/utils/index"
